+++ app/src/main/java/org/simple/clinic/main/TheActivityEffectHandler.kt
@@ -1,0 +1,24 @@
+package org.simple.clinic.main
+
+import com.spotify.mobius.rx2.RxMobius
+import com.squareup.inject.assisted.Assisted
+import com.squareup.inject.assisted.AssistedInject
+import io.reactivex.ObservableTransformer
+import org.simple.clinic.util.scheduler.SchedulersProvider
+
+class TheActivityEffectHandler @AssistedInject constructor(
+    private val schedulers: SchedulersProvider,
+    @Assisted private val uiActions: TheActivityUiActions
+) {
+
+  @AssistedInject.Factory
+  interface InjectionFactory {
+    fun create(uiActions: TheActivityUiActions): TheActivityEffectHandler
+  }
+
+  fun build(): ObservableTransformer<TheActivityEffect, TheActivityEvent> {
+    return RxMobius
+        .subtypeEffectHandler<TheActivityEffect, TheActivityEvent>()
+        .build()
+  }
+}