+++ app/src/main/java/org/simple/clinic/main/TheActivityEvent.kt
@@ -1,0 +1,13 @@
+package org.simple.clinic.main
+
+import org.simple.clinic.widgets.UiEvent
+
+sealed class TheActivityEvent : UiEvent
+
+sealed class LifecycleEvent : TheActivityEvent() {
+  object ActivityStarted : LifecycleEvent()
+  object ActivityStopped : LifecycleEvent()
+  object ActivityDestroyed : LifecycleEvent()
+}
+
+