+++ app/src/main/java/org/simple/clinic/main/TheActivityModel.kt
@@ -1,0 +1,12 @@
+package org.simple.clinic.main
+
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
+
+@Parcelize
+class TheActivityModel : Parcelable {
+
+  companion object {
+    fun create(): TheActivityModel = TheActivityModel()
+  }
+}