+++ app/src/main/java/org/simple/clinic/main/TheActivityUi.kt
@@ -1,0 +1,11 @@
+package org.simple.clinic.main
+
+interface TheActivityUi: TheActivityUiActions {
+  fun showAppLockScreen()
+
+  // This is here because we need to show the same alert in multiple
+  // screens when the user gets verified in the background.
+  fun showUserLoggedOutOnOtherDeviceAlert()
+  fun redirectToLogin()
+  fun showAccessDeniedScreen(fullName: String)
+}