+++ app/src/main/java/org/simple/clinic/main/TheActivityUiActions.kt
@@ -1,0 +1,4 @@
+package org.simple.clinic.main
+
+interface TheActivityUiActions {
+}