+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - Migrated `LoggedOutOfDeviceDialog` to Mobius
 - Setup the UI via an effect in `PatientEntryScreen`
 - Migrated `ConfirmRemoveBloodPressureDialog` to Mobius
-- [In Progress: 11 Aug 2020] Migrate `TheActivity` to Mobius
+- [In Progress: 12 Aug 2020] Migrate `TheActivity` to Mobius
 - [In Progress: 11 Aug 2020] Migrate `UpdatePhoneNumberDialog` to Mobius
 - [In Progress: 11 Aug 2020] Migrate `PatientSearchResultsScreen` to Mobius